d/scripts/steam: Cope with different installation directories
According to Valve, the Steam client is meant to support multiple parallel
installations, with $XDG_DATA_HOME/Steam as the path that is normally used
as the equivalent of C:\Program Files\Steam (referred to as $STEAMDIR
in Valve's scripts). On Linux, ~/.steam/steam ($STEAMCONFIG/steam)
is a symbolic link to the most recently run installation root for
production versions of the Steam client (there can also be a parallel
~/.steam/steambeta for internal beta-test versions), which allows the
Steam client and the Steamworks API to find the Steam installation
root even if the XDG_DATA_HOME environment variable has changed since
installation.

The steam package in Debian has traditionally unpacked the Steam
bootstrapper directly into ~/.steam, which gives that directory a
dual role: the installation root $STEAMDIR, and also the symlink farm
$STEAMCONFIG. Whether to change the default in new installations to
match Valve's is under discussion in #916303.

However, even if we don't want to switch the default, not respecting the
~/.steam/steam and ~/.steam/root symbolic links causes data loss if the
Debian steam package is used as a replacement for Valve's steam-launcher
package: our launcher script would detect that ~/.steam/steam.sh doesn't
exist, and consider that to be a corrupted installation.

Signed-off-by: Simon McVittie <smcv@collabora.com>
Closes: #919467