Demote nvidia-driver-libs-i386 from Recommends to Suggests
These drivers are only required on systems where the NVIDIA proprietary
driver is in use, but the amd64 proprietary driver in nvidia-driver-libs
already Recommends nvidia-driver-libs-i386, and the kernel driver
nvidia-kernel-dkms Recommends nvidia-driver, which Depends on
nvidia-driver-libs; so the only way a user of the proprietary NVIDIA
driver could fail to have nvidia-driver-libs:i386 installed is if
they are already disregarding Recommends.

Keeping this driver in our Suggests means its presence or absence will
normally be mentioned in bug reports for this package.

Signed-off-by: Simon McVittie <smcv@debian.org>