Show a warning if nvidia-driver-libs-i386 needs to be installed
Display a debconf note if the Nvidia proprietary kernel module is loaded,
but the corresponding i386 libraries are not found. This seems to be
the best we can do without adding a Recommends on the i386 libraries,
which would pull in the binary driver even on systems that do not have
Nvidia hardware. (See #918996)

Signed-off-by: Simon McVittie <smcv@debian.org>