Add Conflicts/Replaces on steam-launcher
steam-launcher is a Valve-provided package containing the same
launcher as Debian's steam package, and the same udev rules as Debian's
steam-devices package. They are not co-installable: please install the
steam and steam-devices packages from Debian non-free, *or* the steam
and steam-launcher packages from Valve, but do not mix the two sources.