d/rules: Cope with the upstream release having an epoch
Ubuntu's modified version of this package has an epoch as a result of
an older, unrelated package "sTeam" with a higher version number, which
used to be in Debian and Ubuntu many years ago. Valve's steam package
now has an epoch so that their 1:1.0.0.62 will be compared correctly
against Ubuntu's 1:1.0.0.61.