d/rules: Keep most upstream files in the orig tarball
The upstream "source" tarball has been restructured, and in particular
the udev rules (which are MIT-licensed and come from a separate git
repository) are now in subprojects/steam-devices.

bootstrap*.tar.* contains precompiled libraries without their
corresponding source code (it's available from repo.steampowered.com but
not part of the "source" package), which we intentionally don't ship. The
crashhandler module and the steamerrorreporter executable are also
non-essential binaries.

The other scripts and auxiliary files in the orig tarball are
likely to be useful references, even if we don't actually ship them in
binary packages.