Build-Depend on the libraries Valve put in their bootstrap tarball
This ensures that we have all the necessary metadata for dpkg-shlibdeps
to generate the ${shlibs:Depends} for the proprietary steam executable.

Use SONAMEs for libraries that are unlikely to break ABI, or -dev
packages for libraries that have bumped SONAME in Debian (we're hoping
the proprietary steam executable only depends on them indirectly,
otherwise we would have to bundle a precompiled library or package a
sufficiently old version separately).