Do not keep the launcher script running at startup
Currently, when starting Steam via the Debian launcher, there are three processes active: the Debian launcher shell, the Valve wrapper shell and the actual Steam binary. While the Valve script cannot be controlled and serves the convenience of automatic restarts after internal updates, the Debian launcher has no further purpose, after it started the Valve script.

To reduce process complexity and a bit system memory, this commit starts the Valve script via "exec", so that it replaces the Debian launcher process instead of forking a child.

Signed-off-by: MichaIng <micha@dietpi.com>