Add Recommends on libsdl2-2.0-0:i386
This is an important library for most games and for Steam itself, and
will also pull in up-to-date i386 versions of its dependencies. These
i386 libraries will be used in preference to the rather outdated versions
in the Steam Runtime, mitigating bugs like #990233. Users of wine32 will
have these already.