d/scripts/steam.in: Stop deleting libraries from the Steam Runtime
The Steam Runtime is designed to defer to the libraries from the host
system whenever the version on the host system is newer, or whenever
the versions in the Steam Runtime and on the host system cannot be
compared, so it should not be necessary to edit it in-place.

All situations where this does not work are believed to have been fixed
upstream. If they have not, the upstream developers want to receive
issue reports with diagnostic logs from steam-runtime-system-info, so
that the runtime can be fixed. The beginning of a Debian release cycle
seems like a good time to make this change.