d/scripts/steam.in: Remove workaround for #882607
This shouldn't be necessary any more, particularly since Valve is now
known to be working on a SteamOS 3 update that is based on Arch Linux
(where libdbus assertion failures have always been fatal). The beginning
of a Debian release cycle is a good time to remove old workarounds and
see whether anything breaks.