Repack a version of steam-launcher that is limited to only DFSG files
The .deb never actually contained the full Steam client, only the
"bootstrapper", which is enough to download the rest of Steam and
set it up in the user's home directory.

Since we require Internet access to do that initial setup, we might
as well make use of that Internet access to fetch all the proprietary
files, including the bootstrapper itself. This will make everything in
the .deb DFSG-compliant, which will allow steam and steam-devices to go
into contrib and main, respectively.