Download Valve-provided bootstrapper binary on-demand
This avoids needing to ship it in the .deb, which results in
steam-devices having to be in the non-free archive area as a result of
also being built from a non-free source package.

The icons and .desktop file that come with the proprietary client are
also proprietary, so we substitute a generic "software installation"
icon. During initial setup, we install the proprietary icons into the
user's home directory, where they will take precedence over this
placeholder.