Stop presenting the Steam Subscriber Agreement as a debconf question
Valve's own official .deb doesn't do this, and it really has more to do
with use of a Steam account than it has to do with downloading the client.
The Steam client presents the Steam Subscriber Agreement during creation
of new Steam accounts.

Doing this as a debconf question might have been necessary to comply
with an earlier redistribution license (prior to version 1.0.0.55)
but it doesn't seem to be required any more.