steam-devices: Relax Breaks/Replaces on older steam packages
steam-devices only needs Breaks/Replaces on the steam package if it
shipped its own version of the udev rules that now live in steam-devices.
In Debian, the udev rules moved from steam to steam-devices in 1.0.0.51-1,
while in Valve's packages, all of the actual content (including udev
rules) moved from steam to steam-launcher in 1.0.0.35.

This should make it easier to upgrade steam-installer if the
Architecture: all buildd is not keeping up with amd64/i386.