d/control: Tighten up steam-libs dependencies to match versions
If a user has added Valve's apt repository (but is using our
steam-installer package in preference to Valve's steam-launcher anyway),
we want to get Debian's steam-libs rather than Valve's. Because Debian
is able to target a specific host OS, while Valve is producing generic
packages for multiple OS versions, Debian is able to set tighter
dependencies that are more likely to avoid crashes.